move FileSystemProvider and documentContext to css we can move it out later if it's needed elsewhere
