+++ packages/language-server/src/plugins/documentContext.ts
