+++ packages/language-server/src/svelte-check.ts
@@ -3,7 +3,7 @@ import ts from 'typescript';
 import { Diagnostic, Position, Range } from 'vscode-languageserver';
 import { WorkspaceFolder } from 'vscode-languageserver-protocol';
 import { Document, DocumentManager } from './lib/documents';
-import { FileSystemProvider } from './lib/FileSystemProvider';
+import { FileSystemProvider } from './plugins/css/FileSystemProvider';
 import { Logger } from './logger';
 import { LSConfigManager } from './ls-config';
 import {
